Conflicts:
	plugins/de.cau.cs.kieler.klay.layered/plugin.xml
	plugins/de.cau.cs.kieler.klay.layered/src/de/cau/cs/kieler/klay/layered/intermediate/BigNodesPreProcessor.java
	plugins/de.cau.cs.kieler.klay.layered/src/de/cau/cs/kieler/klay/layered/p2layers/LongestPathLayerer.java
	plugins/de.cau.cs.kieler.klay.layered/src/de/cau/cs/kieler/klay/layered/p2layers/NetworkSimplexLayerer.java
	plugins/de.cau.cs.kieler.klay.layered/src/de/cau/cs/kieler/klay/layered/p5edges/OrthogonalEdgeRouter.java

+++ plugins/de.cau.cs.kieler.klay.layered/src/de/cau/cs/kieler/klay/layered/intermediate/BigNodesPreProcessor.java
@@ -108,7 +108,7 @@ public class BigNodesPreProcessor implements ILayoutProcessor {
         }
 
         // the object spacing in the drawn graph
-        spacing = layeredGraph.getProperty(Properties.OBJ_SPACING);
+        spacing = layeredGraph.getProperty(Properties.OBJ_SPACING).doubleValue();
         direction = layeredGraph.getProperty(LayoutOptions.DIRECTION);
         // the ID for the most recently created dummy node
         dummyID = nodes.size();
